Extract RegisteredTypesSupplier to outer class
Will be required in the CatalogUpgradeScanner that is about to be extracted.
